requirements.txt
alembic==1.12.0
astroid==2.15.6
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dill==0.3.7
dnspython==2.4.2
dominate==2.8.0
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
platformdirs==3.10.0
psycopg2==2.9.9
pylint==2.17.5
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==2.0.4
visitor==0.1.3
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
